requirements.txt
agno>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.36
alembic>=1.13.0
httpx>=0.27.2
openai>=1.54.0
ollama>=0.3.3
mistralai>=1.0.0
psycopg2-binary>=2.9.11
pyyaml>=6.0
python-multipart>=0.0.9
qdrant-client>=1.11.0
